refactor: put the parsing logic in one module
